CSV Parser 2.0.0 (#127)
* Added initial work for internal parser rewrite

* Fixed issue with quoted fields

* Added new tests for fragment stitching

* Added ability to piece together fragments

* New internal CSV parser works

* Refactored CSV guessing to use new parser

* Added some optimizations

* Refactored to use old style loop

* Added some micro-optimizations

* Added some more micro-optimizations

* Simplified switch statement

* Fixed some bugs with quote handling

* Fixed more quoting issues

* Simplified main parse loop

* Update raw_csv_data.cpp

* Added memory mapped IO

* Some more code clean up

* Added dynamic array for storing RawCSVFields

* Most unit tests passing again

* Fixed bug with incomplete fragments

* Fixed read_csv main loop

* CSVRow tests work again

* Update csv_reader_internals.cpp

* Fixed some warnings

* Added more test cases

* Fixed some segfaults

* Updated README and single header

* Updated single header file

* Add reference to cstdlib

* Simplified read_csv. Much faster performance (>500MBps)

* Update csv_reader.cpp

* Modified CSVReader::read_csv

* Switched to std::thread

* Added some optimizations

* Code clean up

* Revert "Code clean up"

This reverts commit 0020ce4735807e5b2b89981b6da9dfc6e4fb79fc.

* More code cleanup

* More code review

* More code cleanup

* More code clean up

* Added a roundtrip test

* Updated docs

* Update test_raw_csv_data.cpp

+++ include/internal/csv_utility.cpp
@@ -54,7 +54,7 @@ namespace csv {
             reader.get_col_names(),
             format.get_delim(),
             reader.size(),
-            (int)reader.get_col_names().size()
+            reader.get_col_names().size()
         };
 
         return info;
